llama-index-integrations/readers/llama-index-readers-chatgpt-conversations/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.codespell]
check-filenames = true
check-hidden = true
# Feel free to un-skip examples, and experimental, you will just need to
# work through many typos (--write-changes and --interactive will help)
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.llamahub]
contains_example = true
import_path = "llama_index.readers.chatgpt_conversations"

[tool.llamahub.class_authors]
ChatGPTConversationsReader = "jasonnathan"
ChatGPTMessageNodeParser = "jasonnathan"

[tool.mypy]
disallow_untyped_defs = true
# Remove venv skip when integrated with pre-commit
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
python_version = "3.8"

[tool.poetry]
authors = ["Jason Joseph Nathan <jjnathanjr@gmail.com>"]
description = "A custom JSON reader for LlamaIndex, parsing ChatGPT-style conversation data."
license = "MIT"
name = "llama-index-readers-chatgpt-conversations"
packages = [{include = "llama_index"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
markdown-it-py = "^3.0.0"
python = ">=3.8.1,<4.0"
llama-index-core = "^0.11.21"
pytest = "^6.2"
pytest-asyncio = "^0.18"
llama-index = "^0.11.21"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "<=23.9.1,>=23.7.0"}
codespell = {extras = ["toml"], version = ">=v2.2.6"}
ipython = "8.10.0"
jupyter = "^1.0.0"
mypy = "0.991"
pre-commit = "3.2.0"
pylint = "2.15.10"
pytest = ">=6.2,<7.0"
pytest-mock = "3.11.1"
ruff = "0.0.292"
tree-sitter-languages = "^1.8.0"
types-Deprecated = ">=0.1.0"
types-PyYAML = "^6.0.12.12"
types-protobuf = "^4.24.0.4"
types-redis = "4.5.5.0"
types-requests = "2.28.11.8"  # TODO: unpin when mypy>0.991
types-setuptools = "67.1.0.0"
